+++ src/world/dungeon.rs
@@ -1,20 +1,11 @@
 use bevy::{prelude::*, render::view::RenderLayers};
-use bevy_proto::prelude::ProtoCommands;
 
-use crate::{
-    player::MovePlayerEvent,
-    world::dimension::{Dimension, SpawnDimension},
-    GameParam, GameState, GAME_HEIGHT,
-};
+use crate::{world::dimension::Dimension, GameState, GAME_HEIGHT};
 
 use super::{
     dimension::{ActiveDimension, DimensionSpawnEvent, EraManager},
-    dungeon_generation::{
-        add_dungeon_chests, add_dungeon_exit_block, gen_new_dungeon, get_player_spawn_tile, Bias,
-    },
-    generation::WorldObjectCache,
-    world_helpers::world_pos_to_tile_pos,
-    TileMapPosition, CHUNK_SIZE,
+    dungeon_generation::{add_dungeon_chests, add_dungeon_exit_block},
+    TileMapPosition,
 };
 
 #[derive(Component)]
@@ -24,13 +15,12 @@ pub struct Dungeon {
 pub struct DungeonPlugin;
 impl Plugin for DungeonPlugin {
     fn build(&self, app: &mut App) {
-        app.add_system(handle_move_player_after_dungeon_gen)
-            .add_systems((
-                add_dungeon_chests,
-                tick_dungeon_timer.run_if(in_state(GameState::Main)),
-                add_dungeon_exit_block,
-                spawn_dungeon_text,
-            ));
+        app.add_systems((
+            add_dungeon_chests,
+            tick_dungeon_timer.run_if(in_state(GameState::Main)),
+            add_dungeon_exit_block,
+            spawn_dungeon_text,
+        ));
     }
 }
 
@@ -42,64 +32,6 @@ pub struct DungeonText;
 
 #[derive(Component, Default)]
 pub struct CachedPlayerPos(pub TileMapPosition);
-
-pub fn spawn_new_dungeon_dimension(
-    game: &mut GameParam,
-    commands: &mut Commands,
-    proto_commands: &mut ProtoCommands,
-    move_player_event: &mut EventWriter<MovePlayerEvent>,
-) {
-    game.clear_dungeon_cache();
-    let player = game.player_query.single().0;
-    let player_pos = game.player().position;
-    commands
-        .entity(player)
-        .insert(CachedPlayerPos(world_pos_to_tile_pos(
-            player_pos.truncate(),
-        )));
-    let grid = gen_new_dungeon(
-        3000,
-        (CHUNK_SIZE * 4 * 2) as usize,
-        Bias {
-            bias: super::dungeon_generation::Direction::Left,
-            strength: 0,
-        },
-    );
-
-    let dim_e = commands
-        .spawn((
-            Dimension,
-            Dungeon { grid: grid.clone() },
-            Dungeontimer(Timer::from_seconds(360., TimerMode::Once)),
-        ))
-        .id();
-    proto_commands.apply("DungeonWorldGenerationParams");
-
-    //cache era data
-    let curr_era = game.era.current_era.clone();
-    game.era
-        .era_generation_cache
-        .insert(curr_era, game.world_obj_cache.clone());
-
-    commands.remove_resource::<WorldObjectCache>();
-    commands.insert_resource(WorldObjectCache::default());
-    commands.entity(dim_e).insert(SpawnDimension);
-
-    if let Some(pos) = get_player_spawn_tile(grid.clone()) {
-        move_player_event.send(MovePlayerEvent { pos });
-    }
-}
-fn handle_move_player_after_dungeon_gen(
-    _new_dungeon: Query<&Dungeon, Added<ActiveDimension>>,
-    _move_player_event: EventWriter<MovePlayerEvent>,
-) {
-    // if let Ok(dungeon) = new_dungeon.get_single() {
-    //     let grid = &dungeon.grid;
-    //     if let Some(pos) = get_player_spawn_tile(grid.clone()) {
-    //         move_player_event.send(MovePlayerEvent { pos });
-    //     }
-    // }
-}
 
 fn tick_dungeon_timer(
     mut commands: Commands,